Refactor: Split MCP JSON classes into separate files
  - Created Models folder to organize MCP protocol classes
  - Moved McpRequest, McpResponse, McpError, McpTool, McpToolResult, McpContent classes
  - Updated namespace and using statements in McpProxy.cs
  - Maintained functionality and successful build
