+++ McpProxy/Models/McpContent.cs
@@ -1,0 +1,12 @@
+using System.Text.Json.Serialization;
+
+namespace CdbMcpServer.Models;
+
+public class McpContent
+{
+    [JsonPropertyName("type")]
+    public string Type { get; set; } = "text";
+    
+    [JsonPropertyName("text")]
+    public string Text { get; set; } = "";
+}
