+++ McpProxy/Models/McpError.cs
@@ -1,0 +1,12 @@
+using System.Text.Json.Serialization;
+
+namespace CdbMcpServer.Models;
+
+public class McpError
+{
+    [JsonPropertyName("code")]
+    public int Code { get; set; }
+
+    [JsonPropertyName("message")]
+    public string Message { get; set; } = "";
+}
