+++ McpProxy/Models/McpRequest.cs
@@ -1,0 +1,19 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace CdbMcpServer.Models;
+
+public class McpRequest
+{
+    [JsonPropertyName("jsonrpc")]
+    public string JsonRpc { get; set; } = "2.0";
+
+    [JsonPropertyName("id")]
+    public int Id { get; set; }
+
+    [JsonPropertyName("method")]
+    public string Method { get; set; } = "";
+
+    [JsonPropertyName("params")]
+    public JsonElement? Params { get; set; }
+}
