+++ McpProxy/Models/McpTool.cs
@@ -1,0 +1,15 @@
+using System.Text.Json.Serialization;
+
+namespace CdbMcpServer.Models;
+
+public class McpTool
+{
+    [JsonPropertyName("name")]
+    public string Name { get; set; } = "";
+
+    [JsonPropertyName("description")]
+    public string Description { get; set; } = "";
+
+    [JsonPropertyName("inputSchema")]
+    public object InputSchema { get; set; } = new();
+}
